DO NOT MERGE: bad hack to try detecting CPU affinity.
This does not work unless the number of possible CPUs is divisible by
8 (the number of bits in a byte) or possibly by 32 or even 64. Kernel
may write zeros in to the mask given to SchedGetaffinity() on places
that are greater than the last enabled CPU, even though those the mask
that has been set (SchedSetaffinity()) would include non-enabled
possible CPUs, too.

The thing is that SchedSetaffinity sets the mask that is remembered by
kernel, but SchedGetaffinity returns currently allowed CPUs instead of
the mask. Therefore, there is no better solution but to set a mask
that covers all possible CPUs.